backend/requirements.txt
django==3.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.11.0
psycopg2==2.8.6
drf-spectacular==0.26.1
django-currentuser==0.5.3
pillow==9.4.0